Move computeTabState out of ScreenPage; drop stale comments/test
- Moved computeTabState from ScreenPage.tsx into useTabExecutionState.ts, clearing a react-refresh/only-export-components lint warning and letting its test drop a renderer-registry mock
- Moved its test to src/hooks/__tests__/useTabExecutionState.computeTabState.test.ts
- Removed a duplicate "defaults busy to false" test in usePageTitle.test.ts
- Reworded comments in useTabExecutionState.ts and ScreenPage.tsx to describe current behavior instead of citing the plan doc or diff-time history
